LabVIEW and MATLAB Files
To facilitate access and collaboration, a new branch named "LabVIEW_MATLAB_Files" has been created. This branch contains the LabVIEW and MATLAB files used in the creation of the control systems and monitoring of the system response. The LabVIEW files provide the graphical programming environment necessary to implement the control algorithms and interface with the electro-hydraulic system. The MATLAB files, on the other hand, offer a powerful computational platform for system analysis, algorithm design, and performance evaluation.

Researchers, engineers, and enthusiasts can access the "LabVIEW_MATLAB_Files" branch to explore and utilize the provided files. These resources enable the replication of the control systems, further experimentation, and analysis of the system response. Leveraging the LabVIEW and MATLAB files enhances transparency, reproducibility, and knowledge sharing within the scientific community.